Changed ints to longs

+++ Prototypes/OrbitSim/src/sim/SolarSystem.java
@@ -17,7 +17,7 @@ public class SolarSystem {
 
 	private ArrayList<SimObject> simObjects;
 	private MainMenu mainMenu;
-	private int timeScale;
+	private long timeScale;
 
 	/**
 	 * @param simObjects
@@ -76,7 +76,7 @@ public class SolarSystem {
 	/**
 	 * @return the timeScale
 	 */
-	public int getTimeScale() {
+	public long getTimeScale() {
 		return timeScale;
 	}
 
@@ -84,7 +84,7 @@ public class SolarSystem {
 	 * @param timeScale
 	 *            the timeScale to set
 	 */
-	public void setTimeScale(int timeScale) {
+	public void setTimeScale(long timeScale) {
 		this.timeScale = timeScale;
 	}
